[refact] RandomEngine 반복 테스트 추가
