+++ src/test/java/racingcargame/domain/RandomEngineTest.java
@@ -4,12 +4,14 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class EngineTest {
+public class RandomEngineTest {
 
     @Test
     @SuppressWarnings("NonAsciiCharacters")
     void 엔진_랜덤_생성_테스트() {
-        Engine engine = Engine.createRandomEngine();
-        assertThat(engine.getPower()).isBetween(0, 9);
+        RandomEngine randomEngine = new RandomEngine();
+        for (int i = 0; i < 100; i++) {
+            assertThat(randomEngine.getPower()).isBetween(0, 9);
+        }
     }
 }
